+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Added property to Cisco Codec VideoComponent to determine if the codec supports video mute based on system software version.
  - Added cameras collection to IConferenceManager
  - Added features for room-wide camera privacy mute
+ 
+### Changed
+ - IConferencePoint implements IPoint<T>
 
 ## [17.2.2] - 2021-04-08
 ### Added